fixed test case and betterer warnings
