+++ src/app/datasets/admin-tab/admin-tab.component.ts
@@ -1,17 +1,17 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { select, Store } from '@ngrx/store';
-import { FileObject } from 'datasets/dataset-details-dashboard/dataset-details-dashboard.component';
-import { Subscription } from 'rxjs';
-import { take } from 'rxjs/operators';
-import { Dataset, Job } from 'shared/sdk';
-import { submitJobAction } from 'state-management/actions/jobs.actions';
-import { getCurrentDatablocks, getCurrentDataset } from 'state-management/selectors/datasets.selectors';
-import { getCurrentUser } from 'state-management/selectors/user.selectors';
+import { Component, OnDestroy, OnInit } from "@angular/core";
+import { select, Store } from "@ngrx/store";
+import { FileObject } from "datasets/dataset-details-dashboard/dataset-details-dashboard.component";
+import { Subscription } from "rxjs";
+import { take } from "rxjs/operators";
+import { Dataset, Job } from "shared/sdk";
+import { submitJobAction } from "state-management/actions/jobs.actions";
+import { getCurrentDatablocks, getCurrentDataset } from "state-management/selectors/datasets.selectors";
+import { getCurrentUser } from "state-management/selectors/user.selectors";
 
 @Component({
-  selector: 'app-admin-tab',
-  templateUrl: './admin-tab.component.html',
-  styleUrls: ['./admin-tab.component.scss']
+  selector: "app-admin-tab",
+  templateUrl: "./admin-tab.component.html",
+  styleUrls: ["./admin-tab.component.scss"]
 })
 export class AdminTabComponent implements OnInit, OnDestroy {
   private subscriptions: Subscription[] = [];
@@ -31,7 +31,7 @@ export class AdminTabComponent implements OnInit, OnDestroy {
   resetDataset() {
     if (confirm("Reset datablocks?")) {
       this.store.pipe(select(getCurrentUser), take(1)).subscribe((user) => {
-        if (user) {
+        if (user && this.dataset) {
           const job = new Job();
           job.emailJobInitiator = user.email;
           job.jobParams = {};
