+++ src/app/datasets/datasets.module.ts
@@ -76,7 +76,7 @@ import { JobEffects } from "state-management/effects/jobs.effects";
 import { LogbookEffects } from "state-management/effects/logbooks.effects";
 import { logbooksReducer } from "state-management/reducers/logbooks.reducer";
 import { DatasetFileUploaderComponent } from "./dataset-file-uploader/dataset-file-uploader.component";
-import { AdminTabComponent } from './admin-tab/admin-tab.component';
+import { AdminTabComponent } from "./admin-tab/admin-tab.component";
 
 @NgModule({
   imports: [
